backend/requirements.txt
fastapi
uvicorn
pydantic
youtube-transcript-api 
openai
python-dotenv